Use evalsha rather than eval
Passing the full script source to redis every time is expensive in redis
compute time and network bandwidth.  loading the script once per store
and calling `evalsha` is faster.  The only tricky part is what to do if
the script cache is tampered with; this requires reloading the script
and a single retry in `compare_and_set_with_ttl`.