Add a method to mark a bucket as overflowed
This is useful when we are running an API client respecting a limiter
that mirrors a remote leaky bucket.  When our client encounters an "out
of quota" exception despite respecting the limiter, the client has more
information than the limiter does and should be able to communicate the
fact that the bucket has overflowed to the rate limit store.